+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # next version
+- application starts with previously selected topic
 - renaming of project and kanban card looks less odd now
 
 # 3.13.0